+++ tests/Database/SelectFields/InterfaceTests/ExampleInterfaceQuery.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Tests\Database\SelectFields\InterfaceTests;
 
+use Closure;
 use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Query;
 use GraphQL\Type\Definition\ResolveInfo;
-use Rebing\GraphQL\Support\SelectFields;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 use Rebing\GraphQL\Tests\Support\Models\Post;
 
@@ -22,10 +22,10 @@ class ExampleInterfaceQuery extends Query
         return Type::listOf(GraphQL::type('ExampleInterface'));
     }
 
-    public function resolve($root, $args, $contxt, ResolveInfo $info, SelectFields $selectFields)
+    public function resolve($root, $args, $contxt, ResolveInfo $info, Closure $getSelectFields)
     {
         return Post
-            ::select($selectFields->getSelect())
+            ::select($getSelectFields()->getSelect())
             ->get();
     }
 }