+++ tests/Support/Objects/ExamplesAuthorizeQuery.php
@@ -7,7 +7,6 @@ namespace Rebing\GraphQL\Tests\Support\Objects;
 use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Query;
 use GraphQL\Type\Definition\ResolveInfo;
-use Rebing\GraphQL\Support\SelectFields;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 
 class ExamplesAuthorizeQuery extends Query
@@ -33,7 +32,7 @@ class ExamplesAuthorizeQuery extends Query
         ];
     }
 
-    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, SelectFields $selectFields)
+    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, Closure $getSelectFields)
     {
         $data = include __DIR__.'/data.php';