+++ tests/Support/Objects/ExamplesFilteredQuery.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Tests\Support\Objects;
 
+use Closure;
 use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Query;
 use GraphQL\Type\Definition\ResolveInfo;
-use Rebing\GraphQL\Support\SelectFields;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 
 class ExamplesFilteredQuery extends Query
@@ -31,7 +31,7 @@ class ExamplesFilteredQuery extends Query
         ];
     }
 
-    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, SelectFields $selectFields): array
+    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, Closure $getSelectFields): array
     {
         $data = include __DIR__.'/data.php';
         $result = [];