+++ tests/Support/Queries/PostsNonNullAndListAndNonNullOfWithSelectFieldsAndModelQuery.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Tests\Support\Queries;
 
+use Closure;
 use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Query;
 use GraphQL\Type\Definition\ResolveInfo;
-use Rebing\GraphQL\Support\SelectFields;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 use Rebing\GraphQL\Tests\Support\Models\Post;
 
@@ -22,10 +22,10 @@ class PostsNonNullAndListAndNonNullOfWithSelectFieldsAndModelQuery extends Query
         return Type::nonNull(Type::listOf(Type::nonNull(GraphQL::type('PostWithModel'))));
     }
 
-    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, SelectFields $selectFields)
+    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, Closure $getSelectFields)
     {
         return Post
-            ::select($selectFields->getSelect())
+            ::select($getSelectFields()->getSelect())
             ->get();
     }
 }